增加了对图片的解码支持，以及添加了TImageBox/TImage class

+++ freeRTOS_TGuiv3/TGui/TPainter/TBufPainter.cpp
@@ -3,7 +3,7 @@
 
 
 
-TBufPainter::TBufPainter(uint32* addr,TRect* rect)
+TBufPainter::TBufPainter(uint8* addr,TRect* rect)
 {
 	font = &Font16x24;
 	textcolor = 0;
